[ELY-441] GSSAPI+GS2 TestSuite (tests optimalization)
